+++ ev.h
@@ -33,6 +33,9 @@
  * ### Features
  * 1. add lua bindings
  * 
+ * ### Bug Fixes
+ * 1. fix: loop may wait infinite with EV_LOOP_MODE_ONCE if there are only endgame events
+ * 
  * 
  * ## v0.0.8 (2023/04/07)
  * 
@@ -201,7 +204,7 @@ extern "C" {
 /**
  * @brief Development version.
  */
-#define EV_VERSION_PREREL           2
+#define EV_VERSION_PREREL           3
 
 /**
  * @brief Version calculate helper macro.