+++ src/timer.h
@@ -17,8 +17,9 @@ EV_LOCAL void ev__init_timer(ev_loop_t* loop);
 /**
  * @brief Process timer.
  * @param[in] loop  Event loop
+ * @return			Active counter
  */
-EV_LOCAL void ev__process_timer(ev_loop_t* loop);
+EV_LOCAL size_t ev__process_timer(ev_loop_t* loop);
 
 #ifdef __cplusplus
 }