+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 ### Features
 1. add lua bindings
 
+### Bug Fixes
+1. fix: loop may wait infinite with EV_LOOP_MODE_ONCE if there are only endgame events
+
 
 ## v0.0.8 (2023/04/07)